scraper/requirements.txt
requests==2.31.0
feedparser==6.0.10
psycopg2-binary==2.9.7
python-dateutil==2.8.2
schedule==1.2.0
sqlalchemy==2.0.21
python-dotenv==1.0.0
lxml==4.9.3
beautifulsoup4==4.12.2
flask==2.3.3